Add software update flyout and UI enhancements
Introduced a new parameter `ShowNewSoftwareUpdateFlyout` in `ContextHelpTitle.razor` to control the visibility of a flyout for software updates. Added the flyout component to `Home.razor` for user notifications. Made minor layout and styling adjustments to `DxButton` and `DxFormLayoutItem`. Also added a `DxMenuItem` reference for the home button in `MainMenu.razor` to improve navigation.